refactor: (react) Remove all circular dependencies.
