+++ packages/react/src/Routing/IncarnateRouteSet.tsx
@@ -1,5 +1,6 @@
 import React, { FC, PropsWithChildren } from 'react';
-import { IncarnateRedirect, IncarnateRoute } from '../.';
+import { IncarnateRedirect } from './IncarnateRedirect';
+import { IncarnateRoute } from './IncarnateRoute';
 
 export type IncarnateRouteSetProps = {
   defaultSubPath?: string;
